+++ lib/tests/service_test_cases.rs
@@ -1,10 +1,12 @@
 use failure::format_err;
 use rudf::model::*;
 use rudf::sparql::{
-    BindingsIterator, GraphPattern, PreparedQuery, QueryOptions, QueryResult, ServiceHandler,
+    BindingsIterator, GraphPattern, QueryOptions, ServiceHandler,
 };
-use rudf::{GraphSyntax, MemoryRepository, Repository, RepositoryConnection, Result};
-use std::io::BufRead;
+use rudf::Result;
+
+mod support;
+use support::*;
 
 #[test]
 fn simple_service_test() {
@@ -214,89 +216,4 @@ fn non_silent_service_test() {
             "This should have been an error since the service fails"
         ),
     }
-}
-
-fn ex(id: String) -> Term {
-    Term::NamedNode(NamedNode::parse(format!("http://example.com/{}", &id)).unwrap())
-}
-
-fn mailto(id: String) -> Term {
-    Term::NamedNode(NamedNode::parse(format!("mailto:{}", &id)).unwrap())
-}
-
-fn literal(str: String) -> Term {
-    Term::Literal(Literal::new_simple_literal(str))
-}
-
-fn make_repository(reader: impl BufRead) -> Result<MemoryRepository> {
-    let repository = MemoryRepository::default();
-    let mut connection = repository.connection()?;
-    connection
-        .load_graph(reader, GraphSyntax::NTriples, None, None)
-        .unwrap();
-    Ok(repository)
-}
-
-fn query_repository<'a>(
-    repository: MemoryRepository,
-    query: String,
-    options: QueryOptions<'a>,
-) -> Result<BindingsIterator<'a>> {
-    match repository
-        .connection()?
-        .prepare_query(&query, options)?
-        .exec()?
-    {
-        QueryResult::Bindings(iterator) => {
-            let (varaibles, iter) = iterator.destruct();
-            let collected = iter.collect::<Vec<_>>();
-            Ok(BindingsIterator::new(
-                varaibles,
-                Box::new(collected.into_iter()),
-            ))
-        }
-        _ => Err(format_err!(
-            "Excpected bindings but got another QueryResult"
-        )),
-    }
-}
-
-fn pattern_repository<'a>(
-    repository: MemoryRepository,
-    pattern: GraphPattern,
-    options: QueryOptions<'a>,
-) -> Result<BindingsIterator<'a>> {
-    match repository
-        .connection()?
-        .prepare_query_from_pattern(&pattern, options)?
-        .exec()?
-    {
-        QueryResult::Bindings(iterator) => {
-            let (varaibles, iter) = iterator.destruct();
-            let collected = iter.collect::<Vec<_>>();
-            Ok(BindingsIterator::new(
-                varaibles,
-                Box::new(collected.into_iter()),
-            ))
-        }
-        _ => Err(format_err!("Expected bindings but got another QueryResult")),
-    }
-}
-
-fn do_query<'a>(
-    reader: impl BufRead,
-    query: String,
-    options: QueryOptions<'a>,
-) -> Result<BindingsIterator<'a>> {
-    let repository = make_repository(reader)?;
-    query_repository(repository, query, options)
-}
-
-fn do_pattern<'a>(
-    reader: impl BufRead,
-    pattern: GraphPattern,
-    options: QueryOptions<'a>,
-) -> Result<BindingsIterator<'a>> {
-    let repository = make_repository(reader)?;
-    pattern_repository(repository, pattern, options)
 }
